components/pdp/requirements.txt
psycopg2-binary==2.9.9
httpx==0.27.0
time
